pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gtdynamics"
version = "1.0.0"
description = "Full kinodynamics constraints for arbitrary robot configurations with factor graphs"
authors = [
    { name = "Varun Agrawal", email = "varunagrawal@gatech.edu" },
    { name = "Frank Dellaert", email = "dellaert@gatech.edu" }
]
readme = "README.md"
license = "BSD-3-Clause"
keywords = ["robotics", "kinematics", "dynamics", "factor graphs"]

classifiers = [
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: BSD License",
]

requires-python = ">=3.10"
dependencies = [
    "numpy>=1.19",
    "gtsam-develop",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/borglab/GTDynamics"
Documentation = "https://github.com/borglab/GTDynamics"
Repository = "https://github.com/borglab/GTDynamics"
"Bug Tracker" = "https://github.com/borglab/GTDynamics/issues"

[tool.pytest.ini_options]
addopts = "--cov=gtdynamics --cov-report html --cov-report term"
norecursedirs = []

[tool.hatch.build]
artifacts = [
    "python/gtdynamics/*.so",
    "python/gtdynamics/*.pyd",
]

[tool.hatch.build.hooks.custom]
path = "hatch_build.py"

[tool.hatch.build.targets.wheel]
packages = ["python/gtdynamics"]

[tool.uv.build-backend]
module-name = "gtdynamics"

[tool.cibuildwheel]
build = "cp312-* cp313-*"
skip = "*-musllinux_* *-win32 pp*"
build-frontend = "pip; args: --no-build-isolation"

[tool.cibuildwheel.linux]
manylinux-x86_64-image = "quay.io/pypa/manylinux_2_34_x86_64"
manylinux-aarch64-image = "quay.io/pypa/manylinux_2_34_aarch64"
before-all = "bash {project}/.github/scripts/cibw_before_all_linux.sh"
before-build = "bash {project}/.github/scripts/cibw_before_build_linux.sh {project}"
repair-wheel-command = "LD_LIBRARY_PATH=/opt/gtdynamics-deps/gtd_current/lib:/opt/gtdynamics-deps/gtsam_current/lib:/opt/gtdynamics-deps/boost/lib:/opt/gtdynamics-deps/sdformat/lib:/opt/gtdynamics-deps/gz-utils/lib:/opt/gtdynamics-deps/gz-math/lib:$LD_LIBRARY_PATH auditwheel repair -w {dest_dir} {wheel} --exclude 'libgtsam*' --exclude 'libmetis-gtsam*' --exclude 'libcephes-gtsam*' --exclude 'libboost_*'"

[tool.cibuildwheel.linux.environment]
INSTALL_PREFIX = "/opt/gtdynamics-deps"
GTSAM_DIR = "/opt/gtdynamics-deps/gtsam_current/lib/cmake/GTSAM"
CMAKE_PREFIX_PATH = "/opt/gtdynamics-deps/gtsam_current:/opt/gtdynamics-deps/gz-cmake4:/opt/gtdynamics-deps/gz-utils:/opt/gtdynamics-deps/gz-math:/opt/gtdynamics-deps/sdformat"
LD_LIBRARY_PATH = "/opt/gtdynamics-deps/gtd_current/lib:/opt/gtdynamics-deps/gtsam_current/lib:/opt/gtdynamics-deps/boost/lib:/opt/gtdynamics-deps/sdformat/lib:/opt/gtdynamics-deps/gz-utils/lib:/opt/gtdynamics-deps/gz-math/lib"
Python_EXECUTABLE = "$(which python)"
PYTHON_EXECUTABLE = "$(which python)"

[tool.cibuildwheel.macos]
before-all = "bash {project}/.github/scripts/cibw_before_all_macos.sh"
before-build = "bash {project}/.github/scripts/cibw_before_build_macos.sh {project}"
repair-wheel-command = "DYLD_LIBRARY_PATH=$HOME/opt/gtdynamics-deps/gtd_current/lib:$HOME/opt/gtdynamics-deps/gtsam_current/lib:$HOME/opt/gtdynamics-deps/boost/lib:$DYLD_LIBRARY_PATH delocate-wheel --require-archs {delocate_archs} -w {dest_dir} -v {wheel} --exclude libgtsam --exclude libmetis-gtsam --exclude libcephes-gtsam --exclude libboost_serialization --exclude libboost_filesystem --exclude libboost_thread --exclude libboost_atomic --exclude libboost_chrono --exclude libboost_date_time --exclude libboost_regex --exclude libboost_timer --exclude libboost_system --exclude libboost_container"

[tool.cibuildwheel.macos.environment]
MACOSX_DEPLOYMENT_TARGET = "$MACOSX_DEPLOYMENT_TARGET"
INSTALL_PREFIX = "$HOME/opt/gtdynamics-deps"
GTSAM_DIR = "$HOME/opt/gtdynamics-deps/gtsam_current/lib/cmake/GTSAM"
CMAKE_PREFIX_PATH = "$HOME/opt/gtdynamics-deps/gtsam_current"
DYLD_LIBRARY_PATH = "$HOME/opt/gtdynamics-deps/gtd_current/lib:$HOME/opt/gtdynamics-deps/gtsam_current/lib:$HOME/opt/gtdynamics-deps/boost/lib"